Use SessionManager.Observer and Session.Obsever for all session related tesets. #3006
This commit applies the new Observer to all session related tests in both normal and private mode.
